Move decorrelator into separate file
